+++ modules/interrogate/interrogate.py
@@ -1,0 +1,4 @@
+def interrogate(image):
+    from modules.interrogate import legacy
+    prompt = legacy.interrogator.interrogate(image)
+    return prompt